+++ web/components/marketplace/app/app-details.tsx
@@ -1,5 +1,0 @@
-export default function AppDetails() {
-  return <div>
-    
-  </div>
-}